'''SONAR''' Added unit tests for SampleDelay, and fixed NUMEROUS bugs.

+++ sandbox/sonar/sonarController/SampleDelay.h
@@ -32,7 +32,7 @@ public:
 private:
 	int buflen;
 	adcdata_t *buf, *bufptr, *bufend;
-	size_t increment;
+	int increment;
 };
